+++ server/buildoutls/commands.py
@@ -1,0 +1,1 @@
+COMMAND_OPEN_PYPI_PAGE = "command/openPypiPage"