+++ server/buildoutls/tests/test_definition.py
@@ -1,6 +1,4 @@
 import pytest
-
-from pygls.server import LanguageServer
 from pygls.lsp.types import (
     Location,
     Position,
@@ -8,6 +6,7 @@ from pygls.lsp.types import (
     TextDocumentIdentifier,
     TextDocumentPositionParams,
 )
+from pygls.server import LanguageServer
 
 from ..server import lsp_definition