+++ server/buildoutls/tests/test_document_link.py
@@ -1,12 +1,11 @@
 import pytest
-
-from pygls.server import LanguageServer
 from pygls.lsp.types import (
     DocumentLinkParams,
     Position,
     Range,
     TextDocumentIdentifier,
 )
+from pygls.server import LanguageServer
 
 from ..server import lsp_document_link